vexriscv: rework vexriscv debug to use a fsm
The vexriscv debug bridge is basically a finite state machine, but
doesn't use any of those primitives.  As a result, it's entirely
synchronous.  This is wasteful, and uses far more flops than is
necessary.

Refactor the debug bridge so that it uses a simple FSM.

This refactor solves an issue where the debug bridge is in a different
clock domain than the CPU itself.

Signed-off-by: Sean Cross <sean@xobs.io>